fix: 중복된 필요없는 코드 제거
